# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LMS Backend API",
    "version": "1.0.0",
    "description": "API documentation cho hệ thống Learning Management System (LMS)",
    "contact": {
      "name": "LMS Team",
      "email": "admin@lms.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.lms.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token để xác thực"
      }
    },
    "schemas": {
...
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer",
            "example": 1
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          },
          "totalUsers": {
            "type": "integer",
            "example": 50
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "example": false
          },
          "limit": {
            "type": "integer",
            "example": 10
          }
        }
      },
...
      "Subject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID duy nhất của môn học",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Tên môn học",
            "example": "Toán học"
          },
          "code": {
            "type": "string",
            "description": "Mã môn học",
            "example": "MATH"
          },
          "description": {
            "type": "string",
            "description": "Mô tả môn học",
            "example": "Môn toán học cơ bản và nâng cao"
          },
          "credits": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "description": "Số tín chỉ",
            "example": 4
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00Z"
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Dữ liệu đầu vào không hợp lệ"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "name"
                },
                "message": {
                  "type": "string",
                  "example": "Tên không được để trống"
                },
                "value": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Không tìm thấy tài nguyên"
          }
        }
      },
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Token không được cung cấp hoặc không hợp lệ"
          }
        }
      },
      "ForbiddenError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Bạn không có quyền truy cập tài nguyên này"
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Lỗi validation dữ liệu đầu vào",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Không tìm thấy tài nguyên",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NotFoundError"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Token không được cung cấp hoặc không hợp lệ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UnauthorizedError"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Không có quyền truy cập",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ForbiddenError"
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/v1/subjects": {
      "get": {
        "summary": "Lấy danh sách tất cả môn học",
        "tags": [
          "Subject Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Trang hiện tại"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Số lượng mỗi trang"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Tìm kiếm theo tên môn học"
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách môn học thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy danh sách môn học thành công"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "subjects": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Subject"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationInfo"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "post": {
        "summary": "Tạo môn học mới",
        "tags": [
          "Subject Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "Toán học"
                  },
                  "code": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "MATH"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Môn toán học cơ bản và nâng cao"
                  },
                  "credits": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10,
                    "example": 4
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo môn học thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo môn học thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Subject"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/v1/subjects/{id}": {
      "get": {
        "summary": "Lấy thông tin chi tiết môn học theo ID",
        "tags": [
          "Subject Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID môn học"
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin môn học thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy thông tin môn học thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Subject"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "put": {
        "summary": "Cập nhật thông tin môn học",
        "tags": [
          "Subject Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID môn học"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "Toán học (Cập nhật)"
                  },
                  "code": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "MATH_UPD"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Môn toán học cơ bản và nâng cao (Cập nhật)"
                  },
                  "credits": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10,
                    "example": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật môn học thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Cập nhật môn học thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Subject"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "delete": {
        "summary": "Xóa môn học (soft delete)",
        "tags": [
          "Subject Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID môn học"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa môn học thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Xóa môn học thành công"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  }
}